requirement.txt
Package            Version
------------------ ---------
annotated-types    0.7.0
anyio              4.9.0
cachetools         5.5.2
certifi            2025.1.31
charset-normalizer 3.4.1
colorama           0.4.6
distro             1.9.0
google-auth        2.38.0
google-genai       1.10.0
h11                0.14.0
httpcore           1.0.7
httpx              0.28.1
idna               3.10
jiter              0.9.0
openai             1.71.0
pip                23.2.1
pyasn1             0.6.1
pyasn1_modules     0.4.2
pydantic           2.11.3
pydantic_core      2.33.1
python-dotenv      1.1.0
regex              2024.11.6
requests           2.32.3
rsa                4.9
sniffio            1.3.1
tiktoken           0.9.0
tqdm               4.67.1
typing_extensions  4.13.1
typing-inspection  0.4.0
urllib3            2.3.0
websockets         15.0.1
